Move Renderer into compose.rendering
