+++ core/src/main/scala/compose/rendering/JsonRenderer.scala
@@ -3,7 +3,6 @@ package compose.rendering
 import java.io.ByteArrayInputStream
 import play.api.libs.json.Writes
 
-import compose.http.Response.Renderer
 import compose.http.Headers
 
 /** Renders any object that's convertible into json into and HTTP response body
